REFACTOR: eliminate long double (#411)
No need to use `long double` -- this is a huge overkill. Floating point
type `double` is far than sufficient for our purpose (even `float` could
be fine, but `double` is a gold standard for most purposes).

---------

+++ source/flow/tcp/swift/swift_cc.cpp
@@ -10,10 +10,9 @@
 
 namespace sim {
 
-TcpSwiftCC::TcpSwiftCC(TimeNs a_base_target, long double a_additive_inc,
-                       long double a_md_beta, long double a_max_mdf,
-                       long double a_fs_range, long double a_fs_min_cwnd,
-                       long double a_fs_max_cwnd)
+TcpSwiftCC::TcpSwiftCC(TimeNs a_base_target, double a_additive_inc,
+                       double a_md_beta, double a_max_mdf, double a_fs_range,
+                       double a_fs_min_cwnd, double a_fs_max_cwnd)
     : m_base_target(a_base_target),
       m_ai(a_additive_inc),
       m_beta_md(a_md_beta),
@@ -21,14 +20,14 @@ TcpSwiftCC::TcpSwiftCC(TimeNs a_base_target, long double a_additive_inc,
       m_fs_range_ns(a_base_target * a_fs_range),
       m_fs_min_cwnd(a_fs_min_cwnd),
       m_fs_max_cwnd(a_fs_max_cwnd),
-      m_cwnd(1.0L),
+      m_cwnd(1.0),
       m_last_decrease(Scheduler::get_instance().get_current_time()),
       m_last_rtt(a_base_target),
       m_retransmit_cnt(0) {
     // ---------- pre‑compute α and β for flow‑scaling ----------
-    const long double inv_sqrt_min = 1.0 / std::sqrt(m_fs_min_cwnd);
-    const long double inv_sqrt_max = 1.0 / std::sqrt(m_fs_max_cwnd);
-    const long double denom = inv_sqrt_min - inv_sqrt_max;
+    const double inv_sqrt_min = 1.0 / std::sqrt(m_fs_min_cwnd);
+    const double inv_sqrt_max = 1.0 / std::sqrt(m_fs_max_cwnd);
+    const double denom = inv_sqrt_min - inv_sqrt_max;
     // α chosen so that flow_term == FS_RANGE at cwnd = FS_MIN_CWND
     m_alpha_flow = m_fs_range_ns / denom;
     m_beta_flow = m_alpha_flow / std::sqrt(m_fs_max_cwnd);
@@ -36,7 +35,7 @@ TcpSwiftCC::TcpSwiftCC(TimeNs a_base_target, long double a_additive_inc,
 
 void TcpSwiftCC::on_ack(TimeNs rtt, [[maybe_unused]] TimeNs avg_rtt,
                         [[maybe_unused]] bool ecn_flag) {
-    if (rtt < TimeNs(1.0L)) {
+    if (rtt < TimeNs(1.0)) {
         throw std::runtime_error("Invalid rtt < 1");
     }
     m_retransmit_cnt = 0;
@@ -46,19 +45,19 @@ void TcpSwiftCC::on_ack(TimeNs rtt, [[maybe_unused]] TimeNs avg_rtt,
     const TimeNs target_delay = compute_target_delay();
 
     const bool can_decrease = compute_can_decreace();
-    long double new_cwnd = m_cwnd;
+    double new_cwnd = m_cwnd;
     // ---------- AIMD ----------
     if (rtt < target_delay) {
         // Additive‑Increase
-        if (new_cwnd > 1.0L)
+        if (new_cwnd > 1.0)
             new_cwnd += (m_ai / new_cwnd);
         else
             new_cwnd += m_ai;
     } else if (can_decrease) {
         // Multiplicative‑Decrease based on «overshoot»
-        const long double overshoot = (rtt - target_delay) / rtt;
-        const long double factor =
-            std::max(1.0L - m_beta_md * overshoot, 1.0L - m_max_mdf);
+        const double overshoot = (rtt - target_delay) / rtt;
+        const double factor =
+            std::max(1. - m_beta_md * overshoot, 1. - m_max_mdf);
         new_cwnd = std::max(M_MIN_CWND, new_cwnd * factor);
     }
 
@@ -68,7 +67,7 @@ void TcpSwiftCC::on_ack(TimeNs rtt, [[maybe_unused]] TimeNs avg_rtt,
 void TcpSwiftCC::on_timeout() {
     const bool can_decrease = compute_can_decreace();
     m_retransmit_cnt++;
-    long double new_cwnd = m_cwnd;
+    double new_cwnd = m_cwnd;
     if (m_retransmit_cnt >= M_RETX_RESET_THRESHOLD) {
         new_cwnd = M_MIN_CWND;
     } else if (can_decrease) {
@@ -78,14 +77,14 @@ void TcpSwiftCC::on_timeout() {
 }
 
 TimeNs TcpSwiftCC::get_pacing_delay() const {
-    if (m_cwnd > 1.0L) {
+    if (m_cwnd > 1.0) {
         return TimeNs(0);
     }
-    long double cwnd_clamped = std::max(m_fs_min_cwnd, m_cwnd);
+    double cwnd_clamped = std::max(m_fs_min_cwnd, m_cwnd);
     return m_last_rtt / cwnd_clamped;
 }
 
-double TcpSwiftCC::get_cwnd() const { return (std::max(m_cwnd, 1.0L)); }
+double TcpSwiftCC::get_cwnd() const { return (std::max(m_cwnd, 1.0)); }
 
 std::string TcpSwiftCC::to_string() const {
     return fmt::format("[target_now: {} ns, cwnd: {:0.3f}]",
@@ -93,9 +92,9 @@ std::string TcpSwiftCC::to_string() const {
 }
 
 TimeNs TcpSwiftCC::compute_target_delay() const {
-    long double cwnd_clamped = std::max(m_fs_min_cwnd, m_cwnd);
+    double cwnd_clamped = std::max(m_fs_min_cwnd, m_cwnd);
     TimeNs flow_term = (m_alpha_flow / std::sqrt(cwnd_clamped)) - m_beta_flow;
-    flow_term = std::clamp(flow_term, TimeNs(0.0L), m_fs_range_ns);
+    flow_term = std::clamp(flow_term, TimeNs(0.0), m_fs_range_ns);
     return m_base_target + flow_term;  // TO-DO: use m_base_target +
                                        // TimeNs(flow_term) + h*hops_num
 }
@@ -105,7 +104,7 @@ bool TcpSwiftCC::compute_can_decreace() const {
     return (current_time - m_last_decrease) > m_last_rtt;
 }
 
-void TcpSwiftCC::update_cwnd(long double new_cwnd) {
+void TcpSwiftCC::update_cwnd(double new_cwnd) {
     new_cwnd = std::clamp(new_cwnd, M_MIN_CWND, M_MAX_CWND);
     if (new_cwnd < m_cwnd) {
         m_last_decrease = Scheduler::get_instance().get_current_time();
